--- src/tamm/tamm.cc
@@ -1,5 +1,0 @@
-#include "tamm/tamm.h"
-
-namespace tamm {
-
-}  // namespace tamm
