+++ src/tamm/tamm.cpp
@@ -1,0 +1,3 @@
+#include "tamm/tamm.hpp"
+
+namespace tamm {} // namespace tamm
