--- src/tamm/tamm.h
@@ -1,19 +1,0 @@
-#ifndef TAMM_TAMM_H_
-#define TAMM_TAMM_H_
-
-#include <array>
-#include <vector>
-#include <cassert>
-#include <memory>
-#include <numeric>
-#include <algorithm>
-#include <cstdlib>
-#include <cmath>
-#include <map>
-#include <iostream>
-#include <type_traits>
-
-#include "tamm/index_space.h"
-
-#endif  // TAMM_TAMM_H_
-
