+++ src/tamm/tamm.hpp
@@ -1,0 +1,18 @@
+#ifndef TAMM_TAMM_HPP
+#define TAMM_TAMM_HPP_
+
+#include <algorithm>
+#include <array>
+#include <cassert>
+#include <cmath>
+#include <cstdlib>
+#include <iostream>
+#include <map>
+#include <memory>
+#include <numeric>
+#include <type_traits>
+#include <vector>
+
+#include "tamm/index_space.hpp"
+
+#endif // TAMM_TAMM_HPP_
